[BACKLOG-14309]  AEL API simplification.
Removed ReportingEvent in favor of consistent event handling
via PDIEvents.
